Move current_epoch from lib to bitcoin.rs
