+++ xmr-btc/src/lib.rs
@@ -60,26 +60,4 @@ pub mod monero;
 pub mod serde;
 pub mod transport;
 
-use crate::bitcoin::{BlockHeight, TransactionBlockHeight, WatchForRawTransaction};
 pub use cross_curve_dleq;
-
-pub async fn current_epoch<W>(
-    bitcoin_wallet: &W,
-    refund_timelock: u32,
-    punish_timelock: u32,
-    lock_tx_id: ::bitcoin::Txid,
-) -> anyhow::Result<Epoch>
-where
-    W: WatchForRawTransaction + TransactionBlockHeight + BlockHeight,
-{
-    let current_block_height = bitcoin_wallet.block_height().await;
-    let t0 = bitcoin_wallet.transaction_block_height(lock_tx_id).await;
-    let t1 = t0 + refund_timelock;
-    let t2 = t1 + punish_timelock;
-
-    match (current_block_height < t1, current_block_height < t2) {
-        (true, _) => Ok(Epoch::T0),
-        (false, true) => Ok(Epoch::T1),
-        (false, false) => Ok(Epoch::T2),
-    }
-}
